requirements.txt
gym
gymnasium
matplotlib
ptan
sklearn-pandas
tensorboardX
torchvision

